Integrate new DatasetMetadata class
