+++ dafni_cli/datasets/dataset_metadata.py
@@ -1,181 +1,243 @@
+from dataclasses import dataclass
+from datetime import datetime
 from io import BytesIO
-from typing import List, Optional, Tuple
+from typing import ClassVar, List, Optional, Tuple
 
 import click
 
 from dafni_cli.api.minio_api import minio_get_request
+from dafni_cli.api.parser import (
+    ParserBaseObject,
+    ParserParam,
+    create_object_from_list_parser,
+    parse_datetime,
+)
 from dafni_cli.api.session import DAFNISession
-from dafni_cli.consts import CONSOLE_WIDTH, DATA_FORMATS, TAB_SPACE
+from dafni_cli.consts import CONSOLE_WIDTH, TAB_SPACE
 from dafni_cli.utils import (
-    check_key_in_dict,
     output_table,
-    process_dict_datetime,
     process_file_size,
     prose_print,
 )
 
 
-class DataFile:
-    """Class to represent the DAFNI Dataset File
-    client model
+@dataclass
+class DataFile(ParserBaseObject):
+    """Dataclass representing a DAFNI dataset file
 
     Methods:
-        __init__(): DataFile constructor
+        download_contents(session: DAFNISession): Downloads the files contents
 
     Attributes:
-        name (str): File Name
-        size (str): File Size
-        format (str): File Format
-        download (str): File Download URL
-        contents (BytesIO): Downloaded File Contents
+        name (str): File name
+        size (str): File size
+        format (str): File format
+        download_url (str): File download url
+        contents (BytesIO): Downloaded file contents (only assigned after
+                            download_contents called)
     """
 
-    def __init__(self, file_dict: Optional[dict] = None):
-        """DataFile Constructor
+    name: str
+    size: str
+    format: str
+    download_url: str
 
-        Args:
-            file_dict (Optional[dict], optional): DataFile Client model to map. Defaults to None.
-        """
-        self.name = None
-        self.size = None
-        self.format = None
-        self.download = None
-        self.contents = None
+    # Separate - only used when actually downloading
+    contents: Optional[BytesIO] = None
 
-        if file_dict:
-            self.set_attributes_from_dict(file_dict)
-
-    def set_attributes_from_dict(self, file_dict: dict):
-        """Function to set the DataFile attributes based on a given
-        Dataset metadata client model
-
-        Args:
-            file_dict (dict): [description]
-        """
-        self.name = check_key_in_dict(file_dict, ["spdx:fileName"])
-        self.size = process_file_size(
-            check_key_in_dict(file_dict, ["dcat:byteSize"], default=None)
-        )
-        self.format = DATA_FORMATS.get(
-            check_key_in_dict(file_dict, ["dcat:mediaType"]), ""
-        )
-        self.download = check_key_in_dict(file_dict, ["dcat:downloadURL"], default=None)
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("name", "spdx:fileName", str),
+        ParserParam("size", "dcat:byteSize", process_file_size),
+        ParserParam("format", "dcat:mediaType", str),
+        ParserParam("download_url", "dcat:downloadURL", str),
+    ]
 
     def download_contents(self, session: DAFNISession):
-        """Function downloads the file using the download url,
-        and saves the contents as a BytesIO object to self.contents
+        """Downloads the file using the download_url and saves the contents as
+        a BytesIO object to self.contents
 
         Args:
             session (DAFNISession): User session
         """
-        self.contents = BytesIO(minio_get_request(self.download, session, content=True))
+        self.contents = BytesIO(
+            minio_get_request(self.download_url, session, content=True)
+        )
 
 
-class DatasetMetadata:
-    """Class to represent the DAFNI Dataset
-    Metadata client model
+@dataclass
+class Creator(ParserBaseObject):
+    type: str
+    name: str
+    id: Optional[str] = None
+    internal_id: Optional[str] = None
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("type", "@type", str),
+        ParserParam("name", "foaf:name", str),
+        ParserParam("id", "@id", str),
+        ParserParam("internal_id", "internalID", str),
+    ]
+
+
+@dataclass
+class Contact(ParserBaseObject):
+    type: str
+    name: str
+    email: str
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("type", "@type", str),
+        ParserParam("name", "vcard:fn", str),
+        ParserParam("email", "vcard:hasEmail", str),
+    ]
+
+    def __str__(self) -> str:
+        """Nicer string representation for printing"""
+        return f"{self.name}, {self.email}"
+
+
+@dataclass
+class Location(ParserBaseObject):
+    id: str
+    type: str
+    label: str
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("id", "@id", str),
+        ParserParam("type", "@type", str),
+        ParserParam("label", "rdfs:label", str),
+    ]
+
+
+@dataclass
+class Publisher(ParserBaseObject):
+    type: str
+    id: Optional[str] = None
+    name: Optional[str] = None
+    internal_id: Optional[str] = None
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("id", "@id", str),
+        ParserParam("type", "@type", str),
+        ParserParam("name", "foaf:name", str),
+        ParserParam("internal_id", "internalID", str),
+    ]
+
+    def __str__(self) -> str:
+        """Nicer string representation for printing"""
+        return f"{self.name}"
+
+
+@dataclass
+class Standard(ParserBaseObject):
+    type: str
+    id: Optional[str] = None
+    label: Optional[str] = None
+
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("id", "@id", str),
+        ParserParam("type", "@type", str),
+        ParserParam("label", "label", str),
+    ]
+
+    def __str__(self) -> str:
+        """Nicer string representation for printing"""
+        return f"{self.label}"
+
+
+@dataclass
+class DatasetMetadata(ParserBaseObject):
+    """Dataclass representing a DAFNI dataset's metadata
 
     Methods:
-        __init__(): DatasetMetadata constructor
-        set_attributes_from_dict(session (DAFNISession), dataset (dict)): Sets the dataset metadata attributes from given client model dict
         output_metadata_details(): Prints key information of the dataset metadata to console.
         output_datafiles_table(): Prints a table to the console of all File related information
         output_metadata_extra_details(): Prints extra details relating to the Dataset Metadata
 
     Attributes:
-        created (str): Date the dataset was created
-        creator (str): Creator of the dataset
-        contact (str): Contact relating to the dataset
+        title (str): Title of the dataset
         description (str): Description of the dataset
+        subject (str): Subject of the dataset
+        created (datetime): Date the dataset was created
+        creators (List[Creator]): List of named creators of the dataset
+        contact (Contact): Contact relating to the dataset
         identifier (str): List of identifiers relating to the dataset
-        location (str): Location the Dataset relates to
-        start_date (str): Dataset start date if applicable
-        end_date (str): Dataset end date if applicable
-        files (List[DataFile]): Files associated with the dataset
+        location (Location): Location the dataset relates to
         keywords (List[str]): Key words relating to the dataset e.g. Transportation
         themes (List[str]): Themes relating to dataset
-        publisher (str): Entity responsible for publishing the dataset
-        issued (str): Date the dataset was issued
-        rights (str): The user rights linked to the Dataset
-        language (str): The langauge used for the dataset
-        standard (str): Any related standards associated
-        update (str): Update Frequency
+        publisher (Publisher): Entity responsible for publishing the dataset
+        issued (datetime): Date and time the dataset was issued
+        language (str): The language used for the dataset
+        standard (Standard): Any related standards associated
+        asset_id (str): Asset identifier for dataset
+        dataset_uuid (str): Dataset UUID
+        version_uuid (str): Version UUID of the latest version of this dataset
+        metadata_uuid (str): Metadata UUID of the latest metadata for this
+                            dataset
+        files (List[DataFile]): Files associated with the dataset
+        rights (List[str] or None): The user rights linked to the Dataset
+        update_frequency (str or None): Update frequency if applicable
+        start_date (datetime or None): Dataset start date if applicable
+        end_date (datetime or None): Dataset end date if applicable
     """
 
-    def __init__(self, dataset_dict: Optional[dict] = None):
-        """DatasetMetadata constructor"""
-        self.created = None
-        self.creator = None
-        self.contact = None
-        self.description = None
-        self.identifier = None
-        self.location = None
-        self.start_date = None
-        self.end_date = None
-        self.files = []
-        self.keywords = []
-        self.themes = None
-        self.publisher = None
-        self.issued = None
-        self.rights = None
-        self.language = None
-        self.standard = None
-        self.update = None
-        self.title = None
-        self.dataset_id = None
-        self.version_id = None
+    title: str
+    description: str
+    subject: str
+    created: datetime
+    creators: List[Creator]
+    contact: Contact
+    identifiers: List[str]
+    location: Location
+    keywords: List[str]
+    themes: List[str]
+    publisher: Publisher
+    issued: datetime
+    language: str
+    standard: Standard
+    asset_id: str
+    dataset_id: str
+    version_id: str
+    metadata_id: str
+    files: List[DataFile]
+    rights: Optional[str] = None
+    update_frequency: Optional[str] = None
+    start_date: Optional[datetime] = None
+    end_date: Optional[datetime] = None
 
-        if dataset_dict:
-            self.set_attributes_from_dict(dataset_dict)
-
-    def set_attributes_from_dict(self, dataset_dict: dict):
-        """Helper function to populate the DatasetMetadata details
-        based on a given DAFNI Dataset metadata client model
-
-        Args:
-            dataset_dict (dict): DAFNI Dataset metadata client model
-        """
-        # Standard Metadata
-        self.created = process_dict_datetime(dataset_dict, ["dct:created"])
-        if check_key_in_dict(dataset_dict, ["dct:creator"], default=None):
-            self.creator = dataset_dict["dct:creator"][0]["foaf:name"]
-        self.contact = check_key_in_dict(
-            dataset_dict, ["dcat:contactPoint", "vcard:hasEmail"]
-        )
-        self.description = check_key_in_dict(dataset_dict, ["dct:description"])
-        self.identifier = check_key_in_dict(dataset_dict, ["dct:identifier"])
-        self.location = check_key_in_dict(dataset_dict, ["dct:spatial", "rdfs:label"])
-        self.start_date = process_dict_datetime(
-            dataset_dict, ["dct:PeriodOfTime", "time:hasBeginning"]
-        )
-        self.end_date = process_dict_datetime(
-            dataset_dict, ["dct:PeriodOfTime", "time:hasEnd"]
-        )
-        self.keywords = check_key_in_dict(dataset_dict, ["dcat:keyword"])
-
-        # Process Files
-        if check_key_in_dict(dataset_dict, ["dcat:distribution"], default=None):
-            self.files = [
-                DataFile(file_dict) for file_dict in dataset_dict["dcat:distribution"]
-            ]
-
-        # Additional Metadata fields
-        self.themes = check_key_in_dict(dataset_dict, ["dcat:theme"])
-        self.publisher = check_key_in_dict(dataset_dict, ["dct:publisher", "foaf:name"])
-        self.issued = process_dict_datetime(dataset_dict, ["dct:issued"])
-        self.rights = check_key_in_dict(dataset_dict, ["dct:rights"])
-        self.language = check_key_in_dict(dataset_dict, ["dct:language"])
-        self.standard = check_key_in_dict(dataset_dict, ["dct:conformsTo", "label"])
-        self.update = check_key_in_dict(dataset_dict, ["dct:accrualPeriodicity"])
-
-        # Version History Fields
-        self.dataset_id = check_key_in_dict(dataset_dict, ["@id", "dataset_uuid"])
-        self.title = check_key_in_dict(dataset_dict, ["dct:title"])
-        self.version_id = check_key_in_dict(dataset_dict, ["@id", "version_uuid"])
+    _parser_params: ClassVar[List[ParserParam]] = [
+        ParserParam("title", "dct:title", str),
+        ParserParam("description", "dct:description", str),
+        ParserParam("subject", "dct:subject", str),
+        ParserParam("created", "dct:created", parse_datetime),
+        ParserParam("creators", "dct:creator", create_object_from_list_parser(Creator)),
+        ParserParam("contact", "dcat:contactPoint", Contact),
+        ParserParam("identifiers", "dct:identifier"),
+        ParserParam("location", "dct:spatial", Location),
+        ParserParam("keywords", "dcat:keyword"),
+        ParserParam("themes", "dcat:theme"),
+        ParserParam("publisher", "dct:publisher", Publisher),
+        ParserParam("issued", "dct:issued", parse_datetime),
+        ParserParam("language", "dct:language", str),
+        ParserParam("standard", "dct:conformsTo", Standard),
+        ParserParam("asset_id", ["@id", "asset_id"], str),
+        ParserParam("dataset_id", ["@id", "dataset_uuid"], str),
+        ParserParam("version_id", ["@id", "version_uuid"], str),
+        ParserParam("metadata_id", ["@id", "metadata_uuid"], str),
+        ParserParam(
+            "files", "dcat:distribution", create_object_from_list_parser(DataFile)
+        ),
+        ParserParam("rights", "dct:rights", str),
+        ParserParam("update_frequency", "dct:accrualPeriodicity", str),
+        ParserParam("end_date", ["dct:PeriodOfTime", "time:hasEnd"], parse_datetime),
+        ParserParam(
+            "start_date", ["dct:PeriodOfTime", "time:hasBeginning"], parse_datetime
+        ),
+    ]
 
     def output_metadata_details(self, long: bool = False):
         """Function to output details relating to the Dataset.
-        The default behavior is to print all standard metadata and
+        The default behaviour is to print all standard metadata and
         a table relating to the associated files.
         If the long option is given, the additional metadata fields are
         also printed to the console
@@ -184,13 +246,15 @@ class DatasetMetadata:
             long (bool, optional): Flag to print additional metadata. Defaults to False.
         """
         click.echo(f"\nCreated: {self.created}")
-        click.echo(f"Creator: {self.creator}")
+        click.echo(f"Creator: {self.creators[0].name}")
         click.echo(f"Contact: {self.contact}")
         click.echo("Description:")
         prose_print(self.description, CONSOLE_WIDTH)
-        click.echo("Identifier: ")
-        prose_print(" ".join(self.identifier), CONSOLE_WIDTH)
-        click.echo(f"Location: {self.location}")
+        click.echo("Identifiers: ")
+        prose_print(
+            " ".join(self.identifiers) if self.identifiers else "None", CONSOLE_WIDTH
+        )
+        click.echo(f"Location: {self.location.label}")
         click.echo(f"Start date: {self.start_date}")
         click.echo(f"End date: {self.end_date}")
         click.echo(f"Key Words:\n {self.keywords}")
@@ -224,17 +288,17 @@ class DatasetMetadata:
         click.echo(f"Publisher: {self.publisher}")
         click.echo(f"Issued: {self.issued}")
         click.echo("Rights:")
-        prose_print(self.rights, CONSOLE_WIDTH)
+        prose_print(self.rights or "None", CONSOLE_WIDTH)
         click.echo(f"Language: {self.language}")
         click.echo(f"Standard: {self.standard}")
-        click.echo(f"Update Frequency: {self.update}")
+        click.echo(f"Update Frequency: {self.update_frequency}")
 
     def output_version_details(self):
         """Prints relevant dataset attributes to command line"""
         click.echo(f"\nTitle: {self.title}")
         click.echo(f"ID: {self.dataset_id}")
         click.echo(f"Version ID: {self.version_id}")
-        click.echo(f"Publisher: {self.publisher}")
+        click.echo(f"Publisher: {self.publisher.name}")
         click.echo(f"From: {self.start_date}{TAB_SPACE}To: {self.end_date}")
         click.echo("Description: ")
         prose_print(self.description, CONSOLE_WIDTH)
